Fixed Scraper Content and Injection
Google had changed the class names for email author and the email body which broke the scraper.

Also changed when the scraper is injected so it would work whenever an email was open, no matter the inbox. Previously, the extension frontent would not load when the page was refreshed on an open email.